MissionProgress component created
